Elm random number generator program.
This simple page is from the Elm getting started guide's introduction to using commands and subscriptions.

Commands and subscriptions are Elm's mechanism for getting "unreliable" data -- typically data that needs to be brought in from some external source; perhaps over HTTP or WebSockets.

To get this type of data in Elm you issue a command or a subscription, which does not guarantee that the call will be succesful (obviously), only that Elm will attempt to fetch the data you requested.